Split relational procedural execution
